refactor: improve api for question condition (#9189)
* refactor: improve api for question condition

* test: ut

* fix: up

* fix: up

* test: ut

* fix: ut

* test: ut

* test: ut

* test: ut

* refactor: ut

* test: ut

* test: ut

* test: ut
