+++ packages/fx-core/package.json
@@ -32,7 +32,8 @@
     "test:botAadAppDriver": "nyc mocha \"tests/component/driver/botAadApp/*.test.ts\"",
     "test:armDriver": "nyc mocha \"tests/component/driver/arm/*.test.ts\"",
     "test:teamsAppDriver": "nyc mocha \"tests/component/driver/teamsApp/*.test.ts\"",
-    "test:coordinator": "nyc mocha \"tests/component/coordinator.test.ts\"",
+    "test:coordinator": "nyc mocha \"tests/component/coordinator/*.test.ts\"",
+    "test:coordinator-create": "nyc mocha \"tests/component/coordinator/coordinator.create.test.ts\"",
     "test:provisionUtil": "nyc mocha \"tests/component/provisionUtils.test.ts\"",
     "test:envUtil": "nyc mocha \"tests/component/envUtil.test.ts\"",
     "test:envMgr": "nyc mocha \"tests/core/environment.test.ts\"",
