+++ packages/fx-core/tests/core/question.test.ts
@@ -406,16 +406,15 @@ describe("updateAadManifestQuestion()", async () => {
     );
     sinon.stub(fs, "pathExists").resolves(true);
     sinon.stub(fs, "readFile").resolves(Buffer.from("${{fake_placeHolder}}"));
-    const res = await validateAadManifestContainsPlaceholder(undefined, inputs);
-    chai.assert.isUndefined(res);
+    const res = await validateAadManifestContainsPlaceholder(inputs);
+    chai.assert.isTrue(res);
   });
   it("validateAadManifestContainsPlaceholder skip", async () => {
     inputs[CoreQuestionNames.AadAppManifestFilePath] = "aadAppManifest";
     sinon.stub(fs, "pathExists").resolves(true);
     sinon.stub(fs, "readFile").resolves(Buffer.from("test"));
-    const res = await validateAadManifestContainsPlaceholder(undefined, inputs);
-    const expectRes = "Skip Current Question";
-    chai.expect(res).to.equal(expectRes);
+    const res = await validateAadManifestContainsPlaceholder(inputs);
+    chai.expect(res).to.equal(false);
   });
   it("getQuestionsForCreateProjectWithoutDotNet for cli", async () => {
     const inputs: Inputs = {
